Add the following new params: HPLC, D15N_NO2+NO3, D18O_NO2+NO3, MICROGELS
HPLC is a placeholder for high performance liquid chromatography
previously known as "high pressure liquid chromatography". These
data are mostly pigments and do not currently go to CCHDO, but we
do need to keep track of where they were measured.

D15N_NO2+NO3 and D18O_NO2+NO3 is the isotopic enrichment of either
15N or 18O in the combined NO2+NO3 chemical species when compared
to a reference standard (VSMOW).

MICROGELS are an abundance (count per vol) of self assembling organic
particles. The units are multiples of 1e6 per l.